Moving get_language_name and get_dict_dir functions from TextSuggest.py to language_layout.py
