+++ TextSuggest.py
@@ -21,6 +21,7 @@ import subprocess as sp
 import sys
 import time
 from collections import Counter
+from language_layout import get_dict_dir, get_language_name
 
 if '--noselect' in sys.argv:
 
@@ -46,41 +47,6 @@ script_cwd = os.path.abspath(os.path.join(__file__, os.pardir))
 
 custom_words_file = os.path.expanduser('~/.Custom_Words.txt')
 
-def get_language_name():
-
-	# This function will return the language name
-	# Reading keyboard layout from shell command
-
-	keyboard_layout = os.popen("setxkbmap -print | awk -F\"+\" '/xkb_symbols/ {print $2}'").read()
-	keyboard_layout = keyboard_layout[:2]
-
-	# Language Layout file contains languages and layouts in
-	# python dictionary format.
-
-	language_layout_file = os.path.join(script_cwd, 'Language_Layout.txt')
-
-	with open(language_layout_file,'r') as f:
-
-		languages = eval(f.read())
-
-	# Language will be detected by layout
-
-	if keyboard_layout in languages:
-
-		return languages[keyboard_layout]
-
-	else:
-
-		return 'English'
-
-def get_dict_dir():
-
-	# Different dictionary for different language
-
-	language = get_language_name()
-
-	return os.path.join(script_cwd, '%sOpenWordList' % language)
-
 def remove_dups(s_list):
 
 	seen = set()
@@ -104,7 +70,7 @@ def get_suggestions(string):
 
 		alphabet = str(current_word[:1])
 
-	dict_dir = get_dict_dir()
+	dict_dir = get_dict_dir(script_cwd)
 
 	dict_file = os.path.join(dict_dir, '%s.txt' % alphabet)
 
@@ -363,4 +329,4 @@ def main(current_word):
 
 if __name__ == '__main__':
 
-	main()
+	main(current_word)
